seperate torch testing
